Changed how {stdin} gets substituted
- The existing `{stdin}` substitution method used the str.format()
method. However, it was causing problems in cases where the user wanted
to input a dict as an arg or kwarg. str.format() tried to parse the
curly brackets as something that needs to be substituted and produced
exceptions as a result
- Changed the substition method to str.replace() which just replaces any
instances of `{stdin}` in the command with the actual stdin value
provided.